create separate package for adapters and move/renamed GroceryListAdapter
