+++ app/src/main/res/layout/layout_grocery_item.xml
@@ -51,6 +51,12 @@
                 android:id="@+id/gUnit"
                 />
 
+            <Spinner
+                android:id="@+id/editGrocery"
+                android:layout_width="wrap_content"
+                android:layout_height="match_parent"
+                android:layout_marginLeft="220dp"/>
+
         </LinearLayout>
 
         <LinearLayout
