+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,9 @@
         <item>lb</item>
         <item>ml</item>
     </string-array>
+    <string-array name="editGroceryActions">
+        <item>Set as shared item</item>
+        <item>Edit Quantity</item>
+        <item>Delete</item>
+    </string-array>
 </resources>
